msi-generation-service/model_api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
torch
timm
numpy
Pillow
python-jose